PDJB-941: simplify flags further
