--- src/main/kotlin/uk/gov/communities/prsdb/webapp/journeys/propertyRegistration/IfEnabledConfig.kt
@@ -1,14 +1,0 @@
-package uk.gov.communities.prsdb.webapp.journeys.propertyRegistration
-
-class IfEnabledConfig<T> {
-    internal var ifEnabledProvider: (() -> T)? = null
-    internal var ifDisabledProvider: (() -> T)? = null
-
-    fun ifEnabled(provider: () -> T) {
-        ifEnabledProvider = provider
-    }
-
-    fun ifDisabled(provider: () -> T) {
-        ifDisabledProvider = provider
-    }
-}
