+++ src/main/kotlin/uk/gov/communities/prsdb/webapp/journeys/propertyRegistration/JointLandlordsPropertyRegistrationStrategy.kt
@@ -49,3 +49,16 @@ class JointLandlordsPropertyRegistrationStrategyImplFlagOn : JointLandlordsPrope
             TaskListItemViewModel.fromTask("registerProperty.taskList.register.inviteJointLandlords", state.jointLandlordsTask),
         )
 }
+
+class IfEnabledConfig<T> {
+    internal var ifEnabledProvider: (() -> T)? = null
+    internal var ifDisabledProvider: (() -> T)? = null
+
+    fun ifEnabled(provider: () -> T) {
+        ifEnabledProvider = provider
+    }
+
+    fun ifDisabled(provider: () -> T) {
+        ifDisabledProvider = provider
+    }
+}
